Amend visit details
Following the implementation of the load tests we quickly realised that
we would benefit from choosing a different prison to test against as the
one initially chosen is used a lot by the team for other testing.

In addition to this we are planning on creating a rake task to delete the
visit requests created via the load test and have changed the visitor
name to be something unique, and therefore making it safer when searching
for the right records to delete.